api/py/requirements.txt
# Python dependencies for Vercel Functions under api/py/
# Mirrors the subset of repo-root requirements.txt the functions need at runtime.

# Web framework (ASGI)
fastapi>=0.100.0
pydantic>=2.0.0

# Environment + HTTP
python-dotenv>=1.0.0
python-dateutil>=2.8.0
requests>=2.28.0
aiohttp>=3.8.0

# LLM
langchain>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
langsmith>=0.0.77

# Data providers
tavily-python>=0.5.0
parallel-web>=0.4.0

# Persistence
supabase>=2.0.0

# YAML for investor profiles
pyyaml>=6.0
